requirements.txt
openai
requests
pymongo
python-dotenv
numpy
certifi
